refactor: split index by RDF/JS spec
